pyproject.toml
[project]
name = "feide-oidc-examples"
version = "0.1.0"
description = "Production-minded Feide OIDC examples (Authorization Code + PKCE, token exchange)."
readme = "README.md"
requires-python = ">=3.13"
license = {file = "LICENSE"}
authors = [{name = "Sikt / Feide"}]
dependencies = [
  "Flask>=3.0",
  "requests>=2.32",
  "python-jose>=3.3.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "black>=24.0",
  "isort>=5.13",
  "basedpyright>=1.20",
  "types-requests>=2.32",
]

[build-system]
requires = ["setuptools>=70"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.black]
line-length = 100
target-version = ["py313"]

[tool.isort]
profile = "black"
line_length = 100
src_paths = ["src", "tests"]

[tool.pyright]
pythonVersion = "3.13"
typeCheckingMode = "strict"
venvPath = "."
venv = ".venv"
include = ["src/feide_login_core", "src/feide_login_full", "src/feide_data_source_api"]
exclude = ["src/feide_login_simple", "tests/test_simple_app.py"]
reportMissingTypeStubs = false
reportUnknownMemberType = true
reportUnknownArgumentType = true
reportUnknownVariableType = true
reportImplicitOverride = true
reportUnusedCallResult = true

executionEnvironments = [
  { root = "src/feide_data_source_api", reportUnusedFunction = "none" }
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"


[tool.setuptools.packages.find]
where = ["src"]
